requirements.txt
openai>=1.3.0
requests
soundfile
pydub
numpy
pyyaml